refactor(pubsub): make exactly once retry loop reusable
